join point -2

+++ algorithm-join-point-2/src/main/java/ru/vood/joinpoint2/infrastructure/flow/mapper/JoinPointDataRowMapper.kt
@@ -23,7 +23,7 @@ class JoinPointDataRowMapper : RowMapper<JoinPointData> {
                 rs.getString(14),
                 rs.getString(15),
                 rs.getInt(16),
-                rs.getInt(17),
+                rs.getLong(17),
 
                 rs.getString(18),
                 rs.getString(19),
